refactor: change response code when checking if user logged in
