+++ backend/controllers/usersController.js
@@ -5,10 +5,19 @@ const IS_PRODUCTION = process.env.NODE_ENV === "production";
 const NUMBER_OF_DAYS = 30;
 
 class UsersController {
-  me(req, res) {
+  async me(req, res) {
+    if (!req.user) {
+      return sendSuccess(res, {
+        message: "No user logged in",
+        data: null,
+      });
+    }
+
+    const loggedInUser = sanitizeUser(req.user);
+
     return sendSuccess(res, {
       message: "User info retrieved",
-      data: sanitizeUser(req.user),
+      data: loggedInUser,
     });
   }
