+++ backend/routes/usersRouter.js
@@ -4,10 +4,13 @@ import { usersController } from "../controllers/usersController.js";
 import { adminRouter } from "./adminRouter.js";
 import { isAdmin } from "../auth/isAdmin.js";
 import { contributionRouter } from "./contributionRouter.js";
-
-usersRouter.post("/logout", usersController.logout);
+import { isAuthenticated } from "../auth/isAuthenticated.js";
 
 usersRouter.get("/me", usersController.me);
+
+usersRouter.use(isAuthenticated);
+
+usersRouter.post("/logout", usersController.logout);
 
 usersRouter.use("/admin", isAdmin, adminRouter);
